requirements.txt
# requirements.txt - Zależności Pythona

```
# Zależności dla LLM Security Test Framework

# Core dependencies
subprocess32>=3.5.4  # Polepszony subprocess
requests>=2.31.0  # HTTP client
click>=8.1.7  # CLI framework

# GUI (opcjonalnie)
pyside6>=6.6.0  # Qt6 dla GUI

# Data processing
pandas>=2.1.4  # Data analysis
numpy>=1.24.0  # Numerics

# Reporting
jinja2>=3.1.2  # Template engine (dla raportów)
markdown2>=2.4.10  # Markdown support

# Testing & Quality
pytest>=7.4.3  # Unit testing
black>=23.12.0  # Code formatter
flake8>=6.1.0  # Linter
pylint>=3.0.3  # Code analysis

# Documentation
sphinx>=7.2.6  # Doc generation
sphinx-rtd-theme>=2.0.0  # RTD theme

# Security & Hashing
pycryptodome>=3.19.0  # Cryptography

# Logging
loguru>=0.7.2  # Better logging

# API & Integration
beautifulsoup4>=4.12.2  # HTML parsing (dla web scraping)
selenium>=4.15.2  # Web automation (dla testów)

# Async support
aiohttp>=3.9.1  # Async HTTP
asyncio>=3.4.3  # Async support

# Utilities
python-dotenv>=1.0.0  # .env file support
colorama>=0.4.6  # Colored terminal output
tqdm>=4.66.1  # Progress bars
```

## Instalacja

```bash
pip install -r requirements.txt
```

## Alternatywnie (minimalny setup bez GUI)

```bash
pip install requests click pandas jinja2 loguru
```
